CPDD and snowfall evolution stored in glacier_evolution.py + small fix
Both CPDD and snowfall are stored in separate csv files in order to track and plot their evolution. A small fix has been done to the way the glacier volume was being computed. The plot scripts have been updated to plot these two new variables. Moreover, the raster files for the DEM and ice thickness for TLT have been added.