Glacier evolution projections processed and stored as netCDF
The glacier projections processing and plotting script now reads and gathers all simulations in a single netCDF file using xarray. Since the dataset is sparse (due to massifs no containing all glaciers), pairing it with the Sparse library would help to reduce the size of the file. The results now can be easily processed, to generate maps or datasets for massifs, glaciers, years, etc...